Implements daily sales and top selling products functionalities

+++ restaurant-pos/src/main/java/com/example/restaurant_pos/model/Product.java
@@ -10,7 +10,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private long id;
 
     private String productName;
     private double price;
@@ -31,11 +31,11 @@ public class Product {
 
     public Product(){}
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getId() {
+    public long getId() {
         return id;
     }
